+++ src/elements/datepicker/datepicker/datepicker.ts
@@ -15,9 +15,6 @@ import type { SbbDatepickerToggleElement } from '../datepicker-toggle.js';
 
 import style from './datepicker.scss?lit&inline';
 
-const FORMAT_DATE =
-  /(^0?[1-9]?|[12]?[0-9]?|3?[01]?)[.,\\/\-\s](0?[1-9]?|1?[0-2]?)?[.,\\/\-\s](\d{1,4}$)?/;
-
 export interface SbbInputUpdateEvent {
   disabled?: boolean;
   readonly?: boolean;
@@ -25,22 +22,26 @@ export interface SbbInputUpdateEvent {
   max?: string | number;
 }
 
+// TODO(breaking-change): Inline deprecated functions in SbbDatepickerElement as public methods
+// where possible and use these methods where the functions are currently used.
+
 /**
  * Given a SbbDatepickerPreviousDayElement, a SbbDatepickerNextDayElement or a SbbDatepickerToggleElement component,
  * it returns the related SbbDatepickerElement reference, if exists.
  * @param element The element potentially connected to the SbbDatepickerElement.
  * @param trigger The id or the reference of the SbbDatePicker.
  */
-export function getDatePicker(
-  element: SbbDatepickerButton | SbbDatepickerToggleElement,
+export function getDatePicker<T = Date>(
+  element: SbbDatepickerButton<T> | SbbDatepickerToggleElement<T>,
   trigger?: string | HTMLElement,
-): SbbDatepickerElement | null | undefined {
+): SbbDatepickerElement<T> | null | undefined {
   if (!trigger) {
-    const parent = element.closest?.('sbb-form-field');
-    return parent?.querySelector('sbb-datepicker');
+    return element
+      .closest?.('sbb-form-field')
+      ?.querySelector<SbbDatepickerElement<T>>('sbb-datepicker');
   }
 
-  return findReferencedElement<SbbDatepickerElement>(trigger);
+  return findReferencedElement<SbbDatepickerElement<T>>(trigger);
 }
 
 /**
@@ -49,13 +50,15 @@ export function getDatePicker(
  * @param delta The number of days to add/subtract from the starting one.
  * @param dateFilter The dateFilter function from the SbbDatepickerElement.
  * @param dateAdapter The adapter class.
+ *
+ * @deprecated Not intended as public API.
  */
-export function getAvailableDate(
-  date: Date,
+export function getAvailableDate<T = Date>(
+  date: T,
   delta: number,
-  dateFilter: ((date: Date) => boolean) | null,
-  dateAdapter: DateAdapter<Date>,
-): Date {
+  dateFilter: ((date: T) => boolean) | null,
+  dateAdapter: DateAdapter<T>,
+): T {
   let availableDate = dateAdapter.addCalendarDays(date, delta);
 
   if (dateFilter) {
@@ -74,13 +77,15 @@ export function getAvailableDate(
  * @param dateFilter The dateFilter function from the SbbDatepickerElement.
  * @param dateAdapter The adapter class.
  * @param min The minimum value to consider in calculations.
+ *
+ * @deprecated Not intended as public API.
  */
-export function findPreviousAvailableDate(
-  date: Date,
-  dateFilter: ((date: Date) => boolean) | null,
-  dateAdapter: DateAdapter<Date>,
+export function findPreviousAvailableDate<T = Date>(
+  date: T,
+  dateFilter: ((date: T) => boolean) | null,
+  dateAdapter: DateAdapter<T>,
   min: string | number | null,
-): Date {
+): T {
   const previousDate = getAvailableDate(date, -1, dateFilter, dateAdapter);
   const dateMin = dateAdapter.deserialize(min);
 
@@ -100,13 +105,15 @@ export function findPreviousAvailableDate(
  * @param dateFilter The dateFilter function from the SbbDatepickerElement.
  * @param dateAdapter The adapter class.
  * @param max The maximum value to consider in calculations.
+ *
+ * @deprecated Not intended as public API.
  */
-export function findNextAvailableDate(
-  date: Date,
-  dateFilter: ((date: Date) => boolean) | null,
-  dateAdapter: DateAdapter<Date>,
+export function findNextAvailableDate<T = Date>(
+  date: T,
+  dateFilter: ((date: T) => boolean) | null,
+  dateAdapter: DateAdapter<T>,
   max: string | number | null,
-): Date {
+): T {
   const nextDate = getAvailableDate(date, 1, dateFilter, dateAdapter);
   const dateMax = dateAdapter.deserialize(max);
 
@@ -126,15 +133,17 @@ export function findNextAvailableDate(
  * @param dateFilter The dateFilter function from the SbbDatepickerElement.
  * @param min The minimum value to consider in calculations.
  * @param max The maximum value to consider in calculations.
+ *
+ * @deprecated Not intended as public API.
  */
-export function isDateAvailable(
-  date: Date,
-  dateFilter: ((date: Date) => boolean) | null,
+export function isDateAvailable<T = Date>(
+  date: T,
+  dateFilter: ((date: T) => boolean) | null,
   min: string | number | null | undefined,
   max: string | number | null | undefined,
 ): boolean {
   // TODO: Get date adapter from config
-  const dateAdapter: DateAdapter<Date> = defaultDateAdapter;
+  const dateAdapter: DateAdapter<T> = readConfig().datetime?.dateAdapter ?? defaultDateAdapter;
   const dateMin = dateAdapter.deserialize(min);
   const dateMax = dateAdapter.deserialize(max);
 
@@ -164,7 +173,7 @@ export const datepickerControlRegisteredEventFactory = (): CustomEvent =>
  * @event {CustomEvent<SbbValidationChangeEvent>} validationChange - Emits whenever the internal validation state changes.
  */
 @customElement('sbb-datepicker')
-export class SbbDatepickerElement extends LitElement {
+export class SbbDatepickerElement<T = Date> extends LitElement {
   public static override styles: CSSResultGroup = style;
   public static readonly events = {
     didChange: 'didChange',
@@ -178,27 +187,48 @@ export class SbbDatepickerElement extends LitElement {
   @property({ type: Boolean }) public wide = false;
 
   /** A function used to filter out dates. */
-  @property({ attribute: 'date-filter' }) public dateFilter: (date: Date | null) => boolean = () =>
-    true;
+  @property({ attribute: false }) public dateFilter: (date: T | null) => boolean = () => true;
 
-  /** A function used to parse string value into dates. */
-  @property({ attribute: 'date-parser' }) public dateParser?: (value: string) => Date | undefined;
+  /**
+   * A function used to parse string value into dates.
+   * @deprecated No longer required.
+   */
+  @property({ attribute: false }) public dateParser?: (value: string) => T | undefined;
 
-  /** A function used to format dates into the preferred string format. */
-  @property() public format?: (date: Date) => string;
+  /**
+   * A function used to format dates into the preferred string format.
+   * @deprecated No longer required.
+   */
+  @property({ attribute: false }) public format?: (date: T) => string;
 
   /** Reference of the native input connected to the datepicker. */
   @property() public input?: string | HTMLElement;
 
+  // TODO: Change undefined to null as a breaking change.
   /** A configured date which acts as the current date instead of the real current date. Recommended for testing purposes. */
   @property()
-  public set now(value: SbbDateLike | undefined) {
+  public set now(value: SbbDateLike<T> | undefined) {
     this._now = this._dateAdapter.getValidDateOrNull(this._dateAdapter.deserialize(value));
   }
-  public get now(): Date {
+  public get now(): T {
     return this._now ?? this._dateAdapter.today();
   }
-  private _now: Date | null = null;
+  private _now?: T | null;
+
+  /** The currently selected date as a Date or custom date provider instance. */
+  @property({ attribute: false })
+  public set valueAsDate(value: SbbDateLike<T> | null) {
+    this._valueAsDate = this._dateAdapter.getValidDateOrNull(this._dateAdapter.deserialize(value));
+    if (this._tryApplyFormatToInput()) {
+      /* Emit blur event when value is changed programmatically to notify
+      frameworks that rely on that event to update form status. */
+      this._inputElement!.dispatchEvent(new Event('blur', { composed: true }));
+    }
+  }
+  public get valueAsDate(): T | null {
+    return this._valueAsDate ?? null;
+  }
+  private _valueAsDate?: T | null;
 
   /**
    * @deprecated only used for React. Will probably be removed once React 19 is available.
@@ -236,137 +266,43 @@ export class SbbDatepickerElement extends LitElement {
     SbbDatepickerElement.events.validationChange,
   );
 
-  @state() private get _inputElement(): HTMLInputElement | null {
-    return this._inputElementState;
-  }
-
-  private set _inputElement(value) {
-    const oldValue = this._inputElementState;
-    this._inputElementState = value;
-    this._registerInputElement(this._inputElementState, oldValue);
-  }
-
-  private _inputElementState: HTMLInputElement | null = null;
-
-  private _findInput(newValue: string | HTMLElement, oldValue: string | HTMLElement): void {
-    if (newValue !== oldValue) {
-      this._inputElement = findInput(this, this.input);
-    }
-  }
-  private _registerInputElement(
-    newValue: HTMLInputElement | null,
-    oldValue: HTMLInputElement | null,
-  ): void {
-    if (newValue !== oldValue) {
-      this._datePickerController?.abort();
-      this._datePickerController = new AbortController();
-
-      if (!this._inputElement) {
-        return;
-      }
-
-      this._inputObserver?.disconnect();
-      this._inputObserver.observe(this._inputElement, {
-        attributeFilter: ['disabled', 'readonly', 'min', 'max', 'value'],
-      });
-
-      this._inputElement.type = 'text';
-
-      if (!this._inputElement.placeholder) {
-        this._inputElement.placeholder = i18nDatePickerPlaceholder[this._language.current];
-      }
-
-      this._inputElement.addEventListener(
-        'change',
-        (event: Event) => {
-          if (!(event instanceof CustomEvent)) {
-            this._valueChanged(event);
-          }
-        },
-        {
-          signal: this._datePickerController.signal,
-        },
-      );
-    }
-  }
-
-  /** Gets the input value with the correct date format. */
-  public getValueAsDate(): Date | undefined {
-    if (this._inputElement && this._inputElement.value) {
-      return this._parse(this._inputElement.value);
-    }
-    return undefined;
-  }
-
-  /** Set the input value to the correctly formatted value. */
-  public setValueAsDate(date: SbbDateLike): void {
-    const parsedDate = date instanceof Date ? date : new Date(date);
-    if (this._inputElement) {
-      this._formatAndUpdateValue(this._inputElement.value, parsedDate);
-      /* Emit blur event when value is changed programmatically to notify
-      frameworks that rely on that event to update form status. */
-      this._inputElement.dispatchEvent(new Event('blur', { composed: true }));
-    }
-  }
-
-  /**
-   * @internal
-   * Whether a custom now is configured.
-   */
-  public hasCustomNow(): boolean {
-    return !!this._now;
-  }
-
-  private _onInputPropertiesChange(mutationsList?: MutationRecord[]): void {
-    this._inputUpdated.emit({
-      disabled: this._inputElement?.disabled,
-      readonly: this._inputElement?.readOnly,
-      min: this._inputElement?.min,
-      max: this._inputElement?.max,
-    });
-
-    if (
-      this._inputElement &&
-      mutationsList &&
-      Array.from(mutationsList).some((e) => e.attributeName === 'value')
-    ) {
-      this._inputElement.value = this._getValidValue(this._inputElement.getAttribute('value')!);
-    }
-  }
+  @state()
+  private _inputElement: HTMLInputElement | null = null;
+  private _inputElementPlaceholderMutable = false;
 
   private _datePickerController!: AbortController;
 
-  private _inputObserver = new AgnosticMutationObserver(this._onInputPropertiesChange.bind(this));
+  private _inputObserver = new AgnosticMutationObserver((mutationsList) => {
+    this._emitInputUpdated();
+    // TODO: Decide whether to remove this logic by adding a value property to the datepicker.
+    if (this._inputElement && mutationsList?.some((e) => e.attributeName === 'value')) {
+      const value = this._inputElement.getAttribute('value');
+      this.valueAsDate = this._dateAdapter.parse(value, this.now) ?? value;
+    }
+  });
 
-  private _dateAdapter: DateAdapter<Date> =
-    readConfig().datetime?.dateAdapter ?? defaultDateAdapter;
+  private _dateAdapter: DateAdapter<T> = readConfig().datetime?.dateAdapter ?? defaultDateAdapter;
 
   private _abort = new SbbConnectedAbortController(this);
   private _language = new SbbLanguageController(this).withHandler(() => {
     if (this._inputElement) {
-      this._inputElement.placeholder = i18nDatePickerPlaceholder[this._language.current];
-      const valueAsDate = this.getValueAsDate();
-      if (valueAsDate) {
-        this._inputElement.value = this._format(valueAsDate);
+      if (this._inputElementPlaceholderMutable) {
+        this._inputElement.placeholder = i18nDatePickerPlaceholder[this._language.current];
+      }
+      if (this.valueAsDate) {
+        this._inputElement.value = this._format(this.valueAsDate);
       }
     }
   });
 
   public override connectedCallback(): void {
     super.connectedCallback();
-    const signal = this._abort.signal;
-    this.addEventListener('datepickerControlRegistered', () => this._onInputPropertiesChange(), {
-      signal,
+    this.addEventListener('datepickerControlRegistered', () => this._emitInputUpdated(), {
+      signal: this._abort.signal,
     });
-    this._inputElement = findInput(this, this.input);
+    this._attachInput();
     if (this._inputElement) {
-      this._inputElement.value = this._getValidValue(this._inputElement.value);
-      this._inputUpdated.emit({
-        disabled: this._inputElement.disabled,
-        readonly: this._inputElement.readOnly,
-        min: this._inputElement.min,
-        max: this._inputElement.max,
-      });
+      this._emitInputUpdated();
     }
   }
 
@@ -374,7 +310,7 @@ export class SbbDatepickerElement extends LitElement {
     super.willUpdate(changedProperties);
 
     if (changedProperties.has('input')) {
-      this._findInput(this.input!, changedProperties.get('input')!);
+      this._attachInput();
     }
     if (
       changedProperties.has('wide') ||
@@ -382,6 +318,9 @@ export class SbbDatepickerElement extends LitElement {
       changedProperties.has('now')
     ) {
       this._datePickerUpdated.emit();
+    }
+    if (changedProperties.has('valueAsDate')) {
+      this._setAriaLiveMessage();
     }
   }
 
@@ -393,111 +332,144 @@ export class SbbDatepickerElement extends LitElement {
 
   protected override firstUpdated(changedProperties: PropertyValues<this>): void {
     super.firstUpdated(changedProperties);
-
-    this._setAriaLiveMessage(this.getValueAsDate());
+    this._setAriaLiveMessage();
   }
 
-  private _parseAndFormatValue(value: string): string {
-    const d = this._parse(value);
-    return !this._dateAdapter.isValid(d) ? value : this._format(d!);
+  /**
+   * Gets the input value with the correct date format.
+   * @deprecated Use property valueAsDate instead.
+   */
+  public getValueAsDate(): T | undefined {
+    return this.valueAsDate ?? undefined;
   }
 
-  private _createAndComposeDate(value: SbbDateLike): string {
-    const date = new Date(value);
-    return this._format(date);
+  /**
+   * Set the input value to the correctly formatted value.
+   * @deprecated Use property valueAsDate instead.
+   */
+  public setValueAsDate(date: SbbDateLike<T>): void {
+    this.valueAsDate = date;
   }
 
-  private _valueChanged(event: Event): void {
-    const value: string = (event.target as HTMLInputElement).value;
-    this._formatAndUpdateValue(value, this._parse(value));
+  /**
+   * @internal
+   * Whether a custom now is configured.
+   */
+  public hasCustomNow(): boolean {
+    return !!this._now;
   }
 
-  /** Applies the correct format to values and triggers event dispatch. */
-  private _formatAndUpdateValue(value: string, valueAsDate: Date | null | undefined): void {
-    if (this._inputElement) {
-      this._inputElement.value = !this._dateAdapter.isValid(valueAsDate)
-        ? value
-        : this._format(valueAsDate!);
+  private _attachInput(): void {
+    const input = findInput(this, this.input);
+    if (this._inputElement === input) {
+      return;
+    } else if (this._inputElement) {
+      this._datePickerController?.abort();
+      this._inputObserver?.disconnect();
+    }
 
-      const isEmptyOrValid =
-        !value ||
-        (!!valueAsDate &&
-          isDateAvailable(
-            valueAsDate,
-            this.dateFilter,
-            this._inputElement?.min,
-            this._inputElement?.max,
-          ));
-      const wasValid = !this._inputElement.hasAttribute('data-sbb-invalid');
-      this._inputElement.toggleAttribute('data-sbb-invalid', !isEmptyOrValid);
-      if (wasValid !== isEmptyOrValid) {
-        this._validationChange.emit({ valid: isEmptyOrValid });
+    this._inputElement = input;
+    if (input) {
+      this._datePickerController = new AbortController();
+      this._inputObserver.observe(input, {
+        attributeFilter: ['disabled', 'readonly', 'min', 'max', 'value'],
+      });
+
+      this._inputElementPlaceholderMutable = !input.placeholder;
+      input.type = 'text';
+      if (this._inputElementPlaceholderMutable) {
+        input.placeholder = i18nDatePickerPlaceholder[this._language.current];
       }
-      this._emitChange(valueAsDate!);
+
+      const options: AddEventListenerOptions = { signal: this._datePickerController.signal };
+      input.addEventListener('input', () => this._parseInput(), options);
+      input.addEventListener('change', () => this._handleInputChange(), options);
+      this._parseInput(true);
+      this._tryApplyFormatToInput();
+      this._validateDate();
     }
   }
 
-  /** Emits the change event. */
-  private _emitChange(date: Date): void {
-    this._setAriaLiveMessage(date);
+  private _emitInputUpdated(): void {
+    const { disabled, readOnly: readonly, min, max } = this._inputElement ?? {};
+    this._inputUpdated.emit({ disabled, readonly, min, max });
+  }
 
+  private _handleInputChange(): void {
+    if (this._tryApplyFormatToInput()) {
+      return;
+    }
+    this._validateDate();
+    this._setAriaLiveMessage();
     this._change.emit();
     this._didChange.emit();
+  }
 
-    if (this._inputElement) {
+  private _tryApplyFormatToInput(): boolean {
+    if (!this._inputElement) {
+      return false;
+    }
+
+    const formattedDate = this.valueAsDate ? this._format(this.valueAsDate!) : '';
+    if (formattedDate && this._inputElement.value !== formattedDate) {
+      this._inputElement.value = formattedDate;
       this._inputElement.dispatchEvent(new InputEvent('input', { bubbles: true, composed: true }));
-      this._inputElement.dispatchEvent(
-        new CustomEvent('change', { bubbles: true, composed: true }),
+      this._inputElement.dispatchEvent(new Event('change', { bubbles: true, composed: true }));
+      return true;
+    }
+
+    return false;
+  }
+
+  private _validateDate(): void {
+    if (!this._inputElement) {
+      return;
+    }
+
+    const isEmptyOrValid =
+      !this._inputElement.value ||
+      (!!this.valueAsDate &&
+        isDateAvailable(
+          this.valueAsDate,
+          this.dateFilter,
+          this._inputElement?.min,
+          this._inputElement?.max,
+        ));
+    const wasValid = !this._inputElement.hasAttribute('data-sbb-invalid');
+    this._inputElement.toggleAttribute('data-sbb-invalid', !isEmptyOrValid);
+    if (wasValid !== isEmptyOrValid) {
+      this._validationChange.emit({ valid: isEmptyOrValid });
+    }
+  }
+
+  private _parseInput(deserializeAsFallback = false): void {
+    const value = this._inputElement!.value;
+    const parse = this.dateParser ?? ((v: string) => this._dateAdapter.parse(v, this.now));
+    // We are assigning directly to the private backing property of valueAsDate
+    // as we don't want to trigger a blur event during this time.
+    this._valueAsDate = this._dateAdapter.getValidDateOrNull(parse(value));
+    if (deserializeAsFallback && !this._valueAsDate) {
+      this._valueAsDate = this._dateAdapter.getValidDateOrNull(
+        this._dateAdapter.deserialize(value),
       );
     }
   }
 
-  private _getValidValue(value: string): string {
-    if (!value) {
-      return '';
-    }
-
-    const match: RegExpMatchArray | null = value.match(FORMAT_DATE);
-
-    if (match?.index === 0) {
-      return this._parseAndFormatValue(value);
-    } else if (Number.isInteger(+value)) {
-      return this._createAndComposeDate(+value);
-    } else if (this._dateAdapter.isValid(new Date(value))) {
-      return this._createAndComposeDate(value);
-    }
-
-    return value;
-  }
-
-  private _parse(value: string): Date | undefined {
-    return this.dateParser ? this.dateParser(value) : this._dateAdapter.parse(value, this.now);
-  }
-
-  private _format(date: Date): string {
+  private _format(date: T): string {
     return this.format ? this.format(date) : this._dateAdapter.format(date);
   }
 
-  private _setAriaLiveMessage(date?: Date): void {
-    const ariaLiveFormatter = new Intl.DateTimeFormat(`${this._language.current}-CH`, {
-      weekday: 'long',
-    });
-
-    const dateFormatter = new Intl.DateTimeFormat('de-CH', {
-      day: '2-digit',
-      month: '2-digit',
-      year: 'numeric',
-    });
-
+  private _setAriaLiveMessage(): void {
     const containerElement: HTMLParagraphElement | null | undefined =
       this.shadowRoot?.querySelector?.<HTMLParagraphElement>('#status-container');
 
-    if (containerElement) {
-      containerElement.innerText = date
-        ? `${i18nDateChangedTo[this._language.current]} ${ariaLiveFormatter.format(
-            date,
-          )}, ${dateFormatter.format(date)}`
-        : '';
+    if (!containerElement) {
+      return;
+    } else if (!this.valueAsDate) {
+      containerElement.innerText = '';
+    } else {
+      const date = this._dateAdapter.format(this.valueAsDate, { weekdayStyle: 'long' });
+      containerElement.innerText = `${i18nDateChangedTo[this._language.current]} ${date}`;
     }
   }